merge these suckers

+++ controllers/api/user-routes.js
@@ -1,6 +1,7 @@
 const router = require('express').Router();
 const bcrypt = require('bcrypt');
 const { User, Garden } = require('../../models');
+const withAuth = require('../../utils/auth');
 
 //Get route for getting all users
 //TODO: remove, potentially?
@@ -13,21 +14,6 @@ router.get('/', async (req, res) => {
     } catch (err) {
         console.log('======\n' + err + '\n======');
         res.status(500).json(err);
-    }
-})
-
-router.get('/signup', (req, res)=>{
-    if (!req.session.user){
-        try{
-        res.status(200).render('createUser');
-        //TODO: frontend code should then take us to the userHome page by setting location.href
-        //to '/dashboard'
-        } catch (err) {
-            console.log('======\n' + err + '\n======');
-            res.status(500).json(err);
-        }
-    } else {
-        res.status(404).end();
     }
 })
 
@@ -50,6 +36,35 @@ router.post('/', async (req, res) => {
     } catch (err) {
         console.log('======\n' + err + '\n======');
         res.status(500).json(err);
+    }
+});
+
+//editing an existing user
+router.put('/:id', withAuth, async (req, res)=>{
+    // TODO: ask if it wouldn't be better to convert req.params.id to a number
+    if(req.session.user && req.session.user.id == req.params.id) {
+        try {
+            const updateUser = await User.update(req.body, { individualHooks: true, where: { id: req.session.user.id }});
+            
+            if (!updateUser) {
+                return res.status(404).json({message: "No user with that id"});
+            }
+
+            const updatedUser = await User.findByPk(req.session.user.id);
+
+            req.session.user = {
+                id: updatedUser.id,
+                username: updatedUser.username,
+                email: updatedUser.email
+            };
+
+            res.status(200).json(updateUser);
+        } catch (err) {
+            console.log('======\n' + err + '\n======');
+            res.status(500).json(err);
+        }
+    } else {
+        res.status(404).json({message: "not found!"})
     }
 });
 
@@ -97,24 +112,24 @@ router.post('/login', async (req, res) => {
 
 //logging out of a currently active account
 router.post('/logout', (req, res) => {
-    try {
-        if (req.session.user) {
+    if (req.session.user) {
+        try {
             req.session.destroy(() => res.status(204).end());
-        } else {
-            res.status(404).end();
+
+        }  catch (err) {
+            console.log('======\n' + err + '\n======');
+            res.status(500).json(err);
         }
-    }  catch (err) {
-        console.log('======\n' + err + '\n======');
-        res.status(500).json(err);
+    } else {
+        res.status(404).end();
     }
 });
 
 //delete existing user's account
 router.delete('/:id', async (req, res) => {
-    //Prevents a user from making a delete request if not logged in
-    if (req.session.user) {
+    //Prevents a user from making a delete request if not logged in to the account to be deleted
+    if (req.session.user && req.session.user.id == req.params.id) {
         try {
-            //Prevents a user from making a delete request to gardens that don't belong to them
             const deleteUser = await User.destroy({
                 where: {
                     id: req.params.id,
@@ -122,7 +137,7 @@ router.delete('/:id', async (req, res) => {
             });
 
             if (!deleteUser) {
-                return res.status(404).json({message: "No user with that id is associated with this user"});
+                return res.status(404).json({message: "No user with that id"});
             }
             req.session.destroy(() => res.status(204).end());
             res.status(200).json(deleteUser);
